[MIPS] Fix lazy-instance template to preserve object alignment.
The alignment attribute of an object was discarded when allocating storage
for the lazy-instance. Alignment is important for MIPS, where doubles
(and objects containing them) must be aligned on 8-byte boundary.

The alignment attributes are included via a macro that is only defined
under GCC, so that this change does not affect other build enviromments.

Change-Id: Ic62c1fc9c5a58f0cf0b4b3c75b56154e3d20aed9
Signed-off-by: Paul Lind <plind@mips.com>